requirements.txt
# 基盤パッケージ（必須）
chromadb==0.5.0
sentence-transformers==2.7.0
google-generativeai==0.8.0
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.0
python-dotenv==1.0.0

# データ処理・可視化

plotly==5.20.0

# テキスト処理最適化
tiktoken==0.7.0
spacy==3.7.0